requirements.txt
Appium-Python-Client==0.24
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
nose==1.3.7
requests==2.13.0
selenium==3.3.1
Werkzeug==0.12.1